refactor(frontend): split step4 report view
